Added js files for browser manipulation. Added file to streamline pulling js files. Moved register tests.
